Reorganizing TRPC helper functions
