+++ app/api/auth/[...nextauth]/adapters/prismaAdapterHashed.ts
@@ -7,7 +7,7 @@ import { AdapterUser } from 'next-auth/adapters';
 import bcrypt from 'bcrypt';
 import { createHash } from 'crypto';
 import { zSiteRoles } from '@/types/sharedZodTypes';
-import { hashEmail } from '@/server/utils/hashedUserHelpers';
+import { hashEmail } from '@/server/utils/userHelpers';
 
 export default function prismaAdapterHashed(prisma: PrismaClient) {
   return {
