+++ server/utils/courseMemberHelpers.ts
@@ -2,7 +2,7 @@ import prisma from '@/prisma';
 import { zCourseRolesType, zSiteRoles } from '@/types/sharedZodTypes';
 import prismaAdapterDefault from '@/app/api/auth/[...nextauth]/adapters/prismaAdapterDefault';
 import { CourseMember } from '@prisma/client';
-import { findDefaultUser } from './defaultUserHelpers';
+import { findUser, hashEmail } from './userHelpers';
 
 /*************************************************************************************
 Search for user with matching course member email 
@@ -23,7 +23,7 @@ export async function createDefaultCourseMember(
   courseMember: CourseMemberInput
 ) {
   const { name, email } = courseMember;
-  const existingUser = await findDefaultUser(email);
+  const existingUser = await findUser(email);
 
   if (!existingUser) {
     await prisma.user.create({
@@ -52,7 +52,7 @@ export type CreateDefaultCourseMemberType = (
 Search for course member with matching user email
 */
 
-export async function findDefaultCourseMember(
+export async function findCourseMember(
   email: string,
   courseId?: string
 ) {
@@ -90,3 +90,54 @@ export type updateDefaultCourseMemberType = (
   id: string,
   memberData: MemberData
 ) => Promise<CourseMember | null>;
+
+/*************************************************************************************
+Search for user with matching course member email (bcrypt)
+If user exists create the courseMember
+If user does not exist create the user and the course member
+*/
+  
+  export async function createHashedCourseMember(
+    courseMember: CourseMemberInput
+  ) {
+    const { name, email } = courseMember;
+    const hashedEmail = hashEmail(email);
+    const existingUser = await findUser(hashedEmail);
+    console.log(existingUser);
+  
+    if (!existingUser) {
+      await prisma.user.create({
+        data: {
+          name: name,
+          email: hashedEmail,
+          role: zSiteRoles.enum.user
+        }
+      });
+    }
+  
+    const resEnrollment = await prisma.courseMember.create({
+      data: {
+        ...courseMember,
+        email: hashedEmail
+      }
+    });
+    return resEnrollment;
+  }
+  
+  export type CreateHashedCourseMemberType = (
+    courseMember: CourseMemberInput
+  ) => Promise<CourseMember>;
+  
+  /*************************************************************************************
+  Search for course member with matching hashed user email
+  */
+  
+  export async function findHashedCourseMember(email: string, courseId?: string) {
+    const where = courseId ? { courseId, email } : { email };
+    return await prisma.courseMember.findFirst({ where });
+  }
+  
+  export type findHashedCourseMemberType = (
+    email: string,
+    courseId?: string
+  ) => Promise<CourseMember | null>;
